split factory module to allow creation of new assets
