+++ experiments/__init__.py
@@ -1,0 +1,3 @@
+from . import dummy
+
+__all__ = ['dummy']
