+++ substratest/factory/__init__.py
@@ -1,0 +1,5 @@
+from .factory import AssetsFactory
+from . import specs
+
+
+__all__ = ['AssetsFactory', 'specs']
